+++ dipu/tests/python/individual_scripts/local_eviron.py
@@ -1,0 +1,1 @@
+../../../torch_dipu/testing/_internal/local_eviron.py
